Add support for "--redirect", to give a 302 on /
I sometimes want to script things using pshs and I don't want to be
forced to know the remote file name.  This patch adds a "--redirect"
option to pshs, which returns a HTTP 302 redirect response instead of
the file list when "/" is requested.  This will cause most clients to
transparently fetch the linked file.  This option only works when there
is a single file passed to pshs, as otherwise the redirect would be
ambiguous.